requirements.txt
requests
customtkinter
pillow
pygame

